+++ src/components/index.js
@@ -14,6 +14,7 @@ export { default as NewOrderModal } from './NewOrderModal';
 export { PaneLoading, ViewLoading } from './Loading';
 export { default as WarningMessage } from './WarningMessage';
 export { default as CalloutRenderer } from './CalloutRenderer';
+export * from './ConnectedTasksJobs';
 export { default as ViewSource } from './ViewSource';
 export { default as AdministrativeNoteList } from './AdministrativeNoteList';
 export { default as FacetOptionFormatter } from './FacetOptionFormatter';